Add profile components
